% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APAP}
\alias{APAP}
\title{Pharmacokinetics of Acetaminophen}
\format{A data frame containing the following columns:
\itemize{
  \item Wt: averaged weight of the study population (kg).
  \item Dose: given oral dose of acetaminophen administered (mg/kg).
  \item Time: time after drug administration (hr).
  \item APAP: concentration of acetaminophen in the sample (mg/L).
  \item AG: concentration of acetaminophen-glucuronide in the sample (mg/L).
  \item AS: concentration of acetaminophen-sulfate in the sample (mg/L).
  \item Study: Sourced reference.
}}
\usage{
APAP
}
\description{
The APAP dataset contains a human experiment of pharmacokinetic data
for acetaminophen, and it's major metabolites.
}
\keyword{datasets}
