% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile}
\alias{compile}
\title{Model Compiler}
\usage{
compile(mName, model = F)
}
\arguments{
\item{mName}{a string giving the name of the model or C file (without extension).}

\item{model}{a logical value to operate the compiler to use model or C file,
the default is set to FALSE to assign the C file in compiling.}
}
\description{
This function is used to compile the model file or C file to generate the executable file in numerical analysis
}
