% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{FFPK}
\alias{FFPK}
\title{Test PK Model for Sensitivity Analysis}
\usage{
FFPK(parameters, time, dose = 1)
}
\arguments{
\item{parameters}{a parameter matrix containing the input sample.}

\item{time}{the given time-points.}

\item{dose}{a given dose.}
}
\description{
The example test model is Flip-flop kinetics (FFPK).
}
