% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_check.R
\name{heat_check}
\alias{heat_check}
\title{Create heatmap}
\usage{
heat_check(x, filter = c("first order", "interaction", "total order"),
  index = "SI", order = F, category = T, text = F)
}
\description{
Create heatmap
}
