% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{pk_model}
\alias{pk_model}
\alias{FFPK}
\title{Example PK Model for Sensitivity Analysis}
\usage{
FFPK(parameters, time, dose = 1)
}
\arguments{
\item{parameters}{a parameter matrix containing the input sample.}

\item{time}{the given time-points.}

\item{dose}{a given dose.}
}
\description{
The example test model is Flip-flop kinetics (FFPK).
\deqn{$C(t) = \frac{F \cdot D \cdot k_a}{(k_a-k_e)V} (e^{-k_et}-e^{-k_at})$/}
}
