% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pksensi-package.R
\docType{package}
\name{pksensi-package}
\alias{pksensi}
\alias{pksensi-package}
\title{pksensi: Global Sensitivity Analysis in Pharmacokinetic Modeling}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Applying the global sensitivity analysis workflow to investigate the parameter uncertainty and sensitivity in pharmacokinetic (PK) models, especially the physiologically-based pharmacokinetic (PBPK) model with multivariate outputs. The package also provide some functions to check the sensitivity measures and its convergence of model parameters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nanhung/pksensi}
  \item Report bugs at \url{https://github.com/nanhung/pksensi/issues}
}

}
\author{
\strong{Maintainer}: Nan-Hung Hsieh \email{nhsieh@cvm.tamu.edu} (0000-0003-0163-2766)

Authors:
\itemize{
  \item Brad Reisfeld
  \item Weihsueh A. Chiu (0000-0002-7575-2368)
}

}
\keyword{internal}
