% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pksensi}
\alias{pksensi}
\alias{pksensi-package}
\title{The \pkg{pksensi} package}
\description{
Reduce parameter dimensionality in pharmacokinetic modeling
and evaluate the robustness of global sensitivity analysis algorithm
}
\details{
The "extended Fourier amplitude sensitivity testing (eFAST)" method,
a variance-based sensitivity analysis method is used to estimate the parameter impact on model output
(Saltelli et al., 1999).
The eFAST is the effective algorithm to determine the influential parameter
in physiologically-based pharmacokinetic model calibration (Hsieh et al., 2018).
}
\references{
A. Saltelli, S. Tarantola and K. Chan, 1999, A quantitative, model independent method for
global sensitivity analysis of model output,
\emph{Technometrics}, 41, 39-56

N-H Hsieh, B Reisfeld, FY Bois, WA, Chiu, 2018, Applying a global sensitivity analysis workflow
to improve the computational efficiencies in physiologically-based pharmacokinetic modeling,
\emph{Front. Pharmacol}, 9, 588.
}
