% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfast99.R
\name{rfast99}
\alias{rfast99}
\title{Extended Fourier Amplitude Sensitivity Test with Random Phase Shift}
\usage{
rfast99(factors, n, M = 4, omega = NULL, q = NULL, q.arg = NULL,
  replicate = 1, conf = 0.95)
}
\arguments{
\item{factors}{an integer for the giving number of factors, or a vector of character strings giving their names.}

\item{n}{an integer for the sampling number.}

\item{M}{an integer specifying the interference parameter.}

\item{omega}{a vector giving the set of frequencies.}

\item{q}{a vector of quantile functions names corresponding to wanted factors distributions.}

\item{q.arg}{a list of quantile functions parameters.}

\item{replicate}{an integer to define the number of replication, the default is 1.}

\item{conf}{the confidence level for replication confidence intervals.}
}
\description{
\code{rfast99} is based on the fast99 function in \pkg{sensitivity} package.
}
