% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_fun.R
\name{solve_fun}
\alias{solve_fun}
\title{Numerical or analytical solution of defined model with parameters}
\usage{
solve_fun(x, times = NULL, parameters, initState, dllname, func, initfunc,
  outnames, method = "lsode", rtol = 1e-08, atol = 1e-12, model = NULL,
  lnparam = F, output)
}
\description{
Numerical or analytical solution of defined model with parameters
}
