% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{heat_check}
\alias{heat_check}
\alias{check}
\alias{print}
\title{Create Heatmap to Detect and Check Parameter Sensitivity}
\usage{
heat_check(x, fit = c("first order", "total order"), vars = NULL,
  times = NULL, SI.cutoff = c(0.01, 0.05), CI.cutoff = c(0.05, 0.1),
  index = "SI", order = F, level = T, text = F)

check(x, times, vars, digits, SI, CI)

print(x, digits)
}
\arguments{
\item{x}{a list of storing information in the defined sensitivity function.}

\item{fit}{a vector of interested output index included \code{first order}, \code{interaction}, and \code{total order}.}

\item{vars}{a logical value or character to specific the display variable in simulation.}

\item{times}{a logical value or character to specific the display time in simulation.}

\item{SI.cutoff}{a vector with two numeric value to specific the cut-off points of sensitivity index in parameter ranking.}

\item{CI.cutoff}{a vector with two numeric value to specific the cut-off points of convergence index in parameter ranking.}

\item{index}{a character to choose sensitivity index \code{SI} (default) or convergence index \code{CI}.}

\item{order}{a logical value indicating whether the parameter should reorder by the value.}

\item{level}{a logical value to use continous or discrete (default) output.}

\item{text}{a logical value to display the calculated indices in the plot.}

\item{digits}{a integer to rounds the values in its first argument.
to the specified number of decimal places (default 4).}

\item{SI}{a numeric value to set the cut-off point for sensitivity index (default 0.1).}

\item{CI}{a numeric vlaue to set the cut-off point for convergence index (default 0.01).}

\item{...}{additional arguments to customize the graphical parameters.}
}
\description{
plot the sensitivity (or convergence) index by heatmap with a given result.
}
\details{
The \code{print} method can used to print sensitivity and convergence indices
with given time-step. The \code{check} method provide the summary of
parameter sensitivity and convergence. The \code{plot} function provide the
time-course functional output of both indices for each parameter.
}
