% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_mcsim}
\alias{install_mcsim}
\title{Solved PK model with given parameters space}
\usage{
install_mcsim(version = "6.0.1", directory = NULL)
}
\arguments{
\item{version}{a character of MCSim version number.}

\item{directory}{a character to assign the directory to install the MCSim.
The default directory is under \code{/home/username} (Linux), \code{/Users/username} (MacOS), and C drive (Windows).}
}
\description{
Download the latest or specific version of MCSim from the official website
\url{https://www.gnu.org/software/mcsim/} and install it to the system directory.
}
