% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_mcsim}
\alias{install_mcsim}
\title{Install MCSim}
\usage{
install_mcsim(version = "6.0.1", directory = NULL, mxstep = 500)
}
\arguments{
\item{version}{a character of MCSim version number.}

\item{directory}{a character to assign the directory to put the MCSim sourced folder.
The default directory is under \code{/home/username} (Linux), \code{/Users/username} (MacOS),
and \code{C:/Users/} (windows).}

\item{mxstep}{a numeric value to assign the maximum number of (internally defined) steps
allowed during one call to the solver (default is 500). The user can increase mxstep to avoid possible error return.}
}
\description{
Download the latest or specific version of MCSim from the official website
\url{https://www.gnu.org/software/mcsim/} and install it to the system directory.
}
\details{
This function aims to help you install MCSim more easily.
However, if you can not install it through this function.
You might need to follow the instruction of MCSim and install it,
manually: \url{https://www.gnu.org/software/mcsim/mcsim.html#Installation}

To install MCSim in Windows, be sure to install Rtools or MinGW first.
}
