% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{mcsim_install}
\alias{mcsim_install}
\alias{mcsim_version}
\title{Install \pkg{GNU MCSim}}
\usage{
mcsim_install(version = "6.2.0", directory = NULL, mxstep = 5000)

mcsim_version()
}
\arguments{
\item{version}{a character of version number.}

\item{directory}{a character to assign the installed directory.}

\item{mxstep}{a numeric value to assign the maximum number of (internally defined) steps
allowed during one call to the solver.}
}
\description{
Download the latest or specific version of \pkg{GNU MCSim} from the official website
(\url{https://www.gnu.org/software/mcsim/}) and install it to the system directory.
}
\details{
This function aims to help users install \pkg{GNU MCSim} more easily.
However, if you can not install it through this function.
The additional way is to follow the instruction and install it manually:
\url{https://www.gnu.org/software/mcsim/mcsim.html#Installation}

The default \code{mxstp} is setting to 5000.
The user can increase \code{mxstp} to avoid possible error return.
If you meet any error when conduct sensitivity analysis,
you can this function to reinstall \pkg{GNU MCSim} and set the higher \code{mxstp}.
The default installed \code{directory} is under \code{/home/username} (Linux),
\code{/Users/username} (MacOS),
and \code{C:/Users/} (windows). To install \pkg{GNU MCSim} in Windows, be sure to install Rtools first.
}
\section{Functions}{
\itemize{
\item \code{mcsim_version}: Return the version number.
}}

\examples{
\dontrun{mcsim_install(version = 6.2.0, mxstep = 10000)}

}
\references{
Bois, F. Y., & Maszle, D. R. (1997).
MCSim: a Monte Carlo simulation program.
\emph{Journal of Statistical Software}, 2(9): 1–60.

\url{https://www.gnu.org/software/mcsim/}
}
