% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pksim.R
\name{pksim}
\alias{pksim}
\title{Pharmacokinetic Simulation from Sampling Parameter}
\usage{
pksim(y, vars = 1, log = F, legend = T, ...)
}
\arguments{
\item{y}{a numeric array created from \code{solve_fun} or \code{solve_mcsim} function.}

\item{vars}{a logical value or character to specific the display variable in simulation (default 1).}

\item{log}{a logical value to transform the y-axis to log scale.}

\item{legend}{a logical value to display the legend in the created plot.}

\item{...}{additional arguments to customize the graphical parameters.}
}
\description{
Pharmacokinetic plot of the output results based on the given parameter (Uncertainty analysis).
If the user define the multiple output in model, the generated result will based on
first model variable (default). A pharmacokinetic plot with median and the range
of min-max, 10%-90%, and 25%-75%.
}
