% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pksim.R
\name{pksim}
\alias{pksim}
\title{PK Simulation from Sampling Parameter}
\usage{
pksim(y, vars = 1, log = F, legend = T, ...)
}
\arguments{
\item{y}{a numeric array created from \code{solve_fun} or \code{solve_mcsim} function.}

\item{vars}{a logical value or character to specific the display variable in simulation (default 1).}

\item{log}{a logical value to transform the y-axis to log scale.}

\item{legend}{a logical value to display the legend in the created plot.}

\item{...}{additional arguments to customize the graphical parameters.}
}
\description{
Create the PK profile of the output results based on the given parameter (uncertainty analysis).
This function is used to visualize the corresponding output PK profile from the input parameters in sensitivity analysis.
If there are multiple outputs in model, the generated result will be the first model variable (default).
A PK plot is generated with median, 25\%-75\%, 10\%-90\%, and minimum-maximum.
}
