% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfast99.R
\name{rfast99}
\alias{rfast99}
\title{Extended Fourier Amplitude Sensitivity Test with Random Phase Shift}
\usage{
rfast99(factors, n, M = 4, omega = NULL, q = NULL, q.arg = NULL,
  replicate = 1, conf = 0.95, ...)
}
\description{
rfast99 implements the so-called "extended-FAST" method
  (Saltelli et al. 1999). This method allows the estimation of first
  order and total Sobol' indices for all the factors.
}
