% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_fun.R
\name{solve_fun}
\alias{solve_fun}
\title{Solve PK model through DeSolve package}
\usage{
solve_fun(x, times = NULL, parameters, initParmsfun = NULL, initState,
  dllname, func, initfunc, outnames, method = "lsode", rtol = 1e-08,
  atol = 1e-12, model = NULL, lnparam = F, output, ...)
}
\arguments{
\item{x}{a list of storing information in the defined sensitivity function.}

\item{times}{a vector to define the given time sequence.}

\item{parameters}{parameters passed to \code{func}.}

\item{initParmsfun}{a character for the given specific initial parameter function.}

\item{initState}{a vector that define the initial values of state variables for the ODE system.}

\item{dllname}{a string giving the name of the shared library (without extension)
that contains the compiled function.}

\item{func}{the name of the function in the dynamically loaded shared library.}

\item{initfunc}{the name of the initialisation function (which initialises values of parameters), as provided in dllname.}

\item{outnames}{the names of output variables calculated in the compiled function \code{func}.}

\item{method}{method used by integrator (default "lsode").}

\item{rtol}{argument passed to integrator (default 1e-8).}

\item{atol}{argument passed to integrator (default 1e-12).}

\item{model}{the defined analytical equation with functional output.}

\item{lnparam}{a logical value that make the statement of the log-transformed parameter (default FALSE).}

\item{output}{a character for the selected output.}

\item{...}{additional arguments for deSolve::ode method.}
}
\description{
The \code{solve_fun} solves for the time-dependent quantity/concentration in different tissues
through the imported deSolve function. It can also solve the function with analytical function.
}
