% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_MCSim.R
\name{solve_MCSim}
\alias{solve_MCSim}
\alias{gen_infile}
\title{Solved PK model through MCSim}
\usage{
solve_MCSim(x, mName, infile.name, outfile.name, parameters = NULL,
  output = NULL, time = NULL, condition = NULL)

gen_infile(infile.name, outfile.name, parameters, output, time, condition,
  rtol = 1e-06, atol = 1e-09)
}
\arguments{
\item{x}{a list of storing information in the defined sensitivity function.}

\item{mName}{a string giving the name of the model or C file (without extension).}

\item{infile.name}{a character to assign the name of in file.}

\item{outfile.name}{a character to assign the name of out file.}

\item{parameters}{a vector to assign the testing parameter.}

\item{output}{a character or a vector to assign the selected output.}

\item{time}{a vector to define the given time sequence.}

\item{condition}{a character to set the specific parameter value in the in file.}

\item{rtol}{argument passed to integrator (default 1e-6).}

\item{atol}{argument passed to integrator (default 1e-9).}
}
\description{
The \code{solve_MCSim} solves for the time-dependent quantity/concentration in different tissues
through the imported deSolve function. It can also solve the function with analytical function.
}
