% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_mcsim.R
\name{solve_mcsim}
\alias{solve_mcsim}
\alias{makemcsim}
\alias{generate_infile}
\title{Solve PK Model Through MCSim}
\usage{
solve_mcsim(x, n = NULL, mName, infile.name, outfile.name,
  setpoint.name = NULL, parameters = NULL, output = NULL, time = NULL,
  condition = NULL)

makemcsim(mName, standalone = F)

generate_infile(infile.name, outfile.name, parameters, output, time, condition,
  rtol = 1e-06, atol = 1e-09, n = NULL, dist = NULL, q.arg = NULL)
}
\arguments{
\item{x}{a list of storing information in the defined sensitivity function.}

\item{n}{a numeric to define the sample number.}

\item{mName}{a string giving the name of the model or C file (without extension).}

\item{infile.name}{a character to assign the name of input file.}

\item{outfile.name}{a character to assign the name of output file.}

\item{setpoint.name}{a character to assign the name of file for parameter matrix.}

\item{parameters}{a character to assign the testing parameters.}

\item{output}{a character or a vector to assign the selected output(s).}

\item{time}{a numeric to define the given time point(s).}

\item{condition}{a character to set the specific parameter value in the input file.}

\item{standalone}{a logic value to create the standalone executable file (default FALSE).}

\item{rtol}{argument passed to integrator (default 1e-6).}

\item{atol}{argument passed to integrator (default 1e-9).}

\item{dist}{a vector of distribution names corresponding to \code{<distribution-name>} in MCSim.}

\item{q.arg}{a list of shape parameters in the sampling distribution (\code{dist}).}
}
\description{
The \code{solve_mcsim} solves for the differential equations of time-dependent quantity/concentration in different tissues/compartments
through MCSim.
The output result is the 4-dimension array with c(model evaluations, replications, time-points, output variables).
}
