% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tell2.R
\name{tell2}
\alias{tell2}
\title{The Decoupling Simulations}
\source{
This function is based on \code{tell} function in \pkg{sensitivity} package,
which is an S3 generic method to estimate sensitivity measures by combining sensitivity object (rfast99) and external simulation results.
}
\usage{
tell2(x, y)
}
\arguments{
\item{x}{a list of storing information in the defined sensitivity function.}

\item{y}{a numeric array generated from the solutions of \code{solve_fun}
or \code{solve_mcsim} function.}
}
\description{
Integrate the decoupling simulations (parameter sequences) and estimation results
to compute the sensitivity measures.
}
